integer division, i love it so much

+++ src/main/java/com/github/nilstrieb/commands/util/EmoteAddCommand.java
@@ -12,6 +12,7 @@ import java.awt.*;
 import java.awt.image.BufferedImage;
 import java.io.*;
 import java.net.URL;
+import java.util.Arrays;
 import java.util.List;
 
 public class EmoteAddCommand extends Command {
@@ -114,8 +115,8 @@ public class EmoteAddCommand extends Command {
     private byte[] resizeImage(byte[] bytes, String format, int size) throws IOException {
         reply("Image size too big (" + bytes.length / 1000 + "kB). Resizing image...");
         Image image = ImageIO.read(new ByteArrayInputStream(bytes));
-        int ratio = image.getHeight(null) / image.getWidth(null);
-        image = image.getScaledInstance(size, size * ratio, Image.SCALE_SMOOTH);
+        double ratio = (double)image.getHeight(null) / (double)image.getWidth(null);
+        image = image.getScaledInstance(size, (int) (size * ratio), Image.SCALE_SMOOTH);
         ByteArrayOutputStream out = new ByteArrayOutputStream();
 
         BufferedImage bufferedImg = new BufferedImage(image.getWidth(null), image.getHeight(null), BufferedImage.TYPE_INT_ARGB);
